# fix `preg_match($pat, $s)` overloading