`$match_all` param was ignored, so `preg_match` behaved like `preg_match_all`;
this led to return value being different from PHP.

```php
preg_match('/\d/', '123');
// => 1 in PHP
// => 3 in KPHP
```

It also led to a degraded performance as we did repeated regexp execution
to find all matches.

Here is a comparison of the old implementation (preg_match like preg_match_all)
and new implementation (that does not try to find all matches):

	name            old time/op  new time/op  delta
	PregMatchShort  1.39µs ± 3%  0.65µs ± 7%  -53.34%  (p=0.000 n=10+8)
	PregMatchLong   47.2µs ± 1%   0.5µs ± 2%  -99.33%  (p=0.000 n=10+9)

Performance results are collected with ktest utility.